Code Drill: methods and method parameters.
At instructor direction I wrote a program to calculate shipping cost based off of some predetermined business rules. The purpose of this exercise was to familiarize myself with methods and their usage in c# code. This was also a valuable learning experience because I had to refactor my original code due to input errors. by doing this I gained valuable insight into how and especially when it is useful to implement method parameters.